pyproject.toml
[project]
name = 'vikunja-dump'
version = '0.0.1'
description = 'Dump text from Vikunja tasks to Markdown files'
authors = [{ name = 'Jordan Cook' }]
license = 'MIT'
rreadme = 'README.md'
requires-python = '>=3.10'
dependencies = [
    'python-dateutil',
    'python-dotenv',
    'requests',
    'html2text',
]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 100
output-format = 'grouped'
target-version = 'py310'

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.lint]
select = ['B', 'C4', 'C90', 'E', 'F', 'I']
